+++ src/kernel/portable.rs
@@ -33,8 +33,8 @@ static mut ULL_NEXT_TIME: u64 = 0;
 pub const ULL_MACHINE_TIMER_COMPARE_REGISTER_BASE: UBaseType = CONFIG_MTIMECMP_BASE_ADDRESS;
 
 extern "C" {
-    pub static mut pullMachineTimerCompareRegister: u32;
-    pub static mut pullNextTime: u32;
+    pub static mut pullMachineTimerCompareRegister: usize;
+    pub static mut pullNextTime: usize;
     pub static mut xISRStackTop: *const StackType;
 }
 //todo:safe  global var and pointer
@@ -62,7 +62,7 @@ pub fn v_port_setup_timer_interrupt() {
     let mut ul_current_time_high: UBaseType;
 
     unsafe {
-        pullNextTime = &ULL_NEXT_TIME as *const u64 as u32;
+        pullNextTime = &ULL_NEXT_TIME as *const u64 as usize;
         uxTimerIncrementsForOneTick = CONFIG_CPU_CLOCK_HZ / CONFIG_TICK_RATE_HZ;
         asm!("csrr {0}, mhartid",out(reg) ul_hart_id);
         pullMachineTimerCompareRegister = ULL_MACHINE_TIMER_COMPARE_REGISTER_BASE + ul_hart_id * 4;
@@ -87,17 +87,17 @@ pub fn v_port_setup_timer_interrupt() {
 pub fn auto_set_currentTcb() {
     unsafe {
         match get_current_tcb(){
-            Some(x) => pxCurrentTCB = x as *const tskTaskControlBlock as u32,
+            Some(x) => pxCurrentTCB = x as *const tskTaskControlBlock as usize,
             None => pxCurrentTCB = 0,
         }
     }
 }
 pub fn x_port_start_scheduler() -> bool {
     unsafe {
-        xISRStackTop = (&X_ISRSTACK[CONFIG_ISR_STACK_SIZE_WORDS - 1]) as *const u32;
+        xISRStackTop = (&X_ISRSTACK[CONFIG_ISR_STACK_SIZE_WORDS - 1]) as *const usize;
     }
     v_port_setup_timer_interrupt();
-    let mut tmp: u32 = 0x800;
+    let mut tmp: usize = 0x800;
     if CONFIG_MTIME_BASE_ADDRESS != 0 && CONFIG_MTIMECMP_BASE_ADDRESS != 0 {
         tmp = 0x880;
     }
